requirements.txt
accelerate
aioboto3
aiohttp
numpy

# database packages
asyncpg
graspologic
hnswlib
nano-vectordb
neo4j
networkx
ollama
openai
oracledb
pyvis
tenacity
# lmdeploy[all]

# LLM packages
tiktoken
torch
transformers
xxhash

# server packages
fastapi
uvicorn
python-multipart
nest_asyncio

# Testing packages
pytest
pytest-cov
pytest-vcr
pytest-asyncio
vcrpy
future
pytest-xdist
